requirements.txt
numpy==1.24.2
pillow==9.4.0
scikit-image==0.21.0
opencv-python-headless==4.7.0.72
torch==1.13.0+cu116
onnxruntime-gpu==1.14.0
openslide-python==1.3.0
tqdm==4.65.0
pandas==2.2.0
--extra-index-url https://download.pytorch.org/whl/cu116